requirements.txt
# Необходимые библиотеки для AI-агента

# Основной фреймворк для создания агентов
smolagents==1.12.0
smolagents[openai]==1.12.0

# Для работы с переменными окружения
python-dotenv==1.0.0

# Дополнительные зависимости (если потребуются)
requests>=2.32.3
